Fix tests failure
