+++ src/Ztm.Zcoin.Rpc.Tests/RpcClientTests.cs
@@ -1,76 +1,19 @@
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Moq;
 using NBitcoin;
 using NBitcoin.RPC;
-using NBitcoin.Tests;
 using Xunit;
 using Ztm.Zcoin.NBitcoin;
 using Ztm.Zcoin.NBitcoin.Exodus;
-using Ztm.Zcoin.Testing;
 
 namespace Ztm.Zcoin.Rpc.Tests
 {
-    public class RpcClientTests : IDisposable
+    public sealed class RpcClientTests : RpcClientTesting
     {
-        static readonly IEnumerable<ITransactionPayloadEncoder> ExodusEncoders = new[]
-        {
-            new SimpleSendEncoder()
-        };
-
-        readonly NodeBuilder nodes;
-        readonly Lazy<RpcClient> subject;
-        bool disposed;
-
-        public RpcClientTests()
-        {
-            this.nodes = NodeBuilderFactory.CreateNodeBuilder(GetType());
-
-            try
-            {
-                this.nodes.ConfigParameters.Add("autocommit", "0");
-                this.nodes.ConfigParameters.Add("dandelion", "0");
-                this.nodes.ConfigParameters.Add("exodus", "1");
-
-                Node = this.nodes.CreateNode(true);
-                Client = Node.CreateRPCClient();
-
-                Factory = new RpcFactory(
-                    this.nodes.Network,
-                    Node.RPCUri,
-                    RPCCredentialString.Parse(Node.GetRPCAuth()),
-                    new TransactionEncoder(ExodusEncoders)
-                );
-
-                this.subject = new Lazy<RpcClient>(CreateSubject, LazyThreadSafetyMode.ExecutionAndPublication);
-            }
-            catch
-            {
-                this.nodes.Dispose();
-                throw;
-            }
-        }
-
-        protected RPCClient Client { get; }
-
-        protected RpcFactory Factory { get; }
-
-        protected Network Network => this.nodes.Network;
-
-        protected CoreNode Node { get; }
-
-        protected RpcClient Subject => this.subject.Value;
-
-        FakeRpcClient FakeSubject => (FakeRpcClient)Subject;
-
-        public void Dispose()
-        {
-            Dispose(true);
-            GC.SuppressFinalize(this);
-        }
+        new FakeRpcClient Subject => (FakeRpcClient)base.Subject;
 
         [Fact]
         public void Constructor_WithNullFactory_ShouldThrow()
@@ -89,17 +32,17 @@ namespace Ztm.Zcoin.Rpc.Tests
         [Fact]
         public void Constructor_WhenSuccess_ShouldInitializeProperties()
         {
-            Assert.Same(Client, FakeSubject.Client);
-            Assert.Same(Factory, FakeSubject.Factory);
+            Assert.Same(Client, Subject.Client);
+            Assert.Same(Factory, Subject.Factory);
         }
 
         [Fact]
         public void Dispose_WhenInvoke_ShouldInvokeProtectedVersion()
         {
-            FakeSubject.Dispose();
-            FakeSubject.Dispose();
+            Subject.Dispose();
+            Subject.Dispose();
 
-            FakeSubject.StubbedDispose.Verify(
+            Subject.StubbedDispose.Verify(
                 f => f(true),
                 Times.Exactly(2)
             );
@@ -158,7 +101,7 @@ namespace Ztm.Zcoin.Rpc.Tests
         {
             return Assert.ThrowsAsync<ArgumentNullException>(
                 "tx",
-                () => FakeSubject.PopulateExodusInformationAsync(null, CancellationToken.None)
+                () => Subject.PopulateExodusInformationAsync(null, CancellationToken.None)
             );
         }
 
@@ -167,7 +110,7 @@ namespace Ztm.Zcoin.Rpc.Tests
         {
             foreach (var tx in Network.GetGenesis().Transactions)
             {
-                await FakeSubject.PopulateExodusInformationAsync(tx, CancellationToken.None);
+                await Subject.PopulateExodusInformationAsync(tx, CancellationToken.None);
 
                 Assert.Null(tx.GetExodusTransaction());
             }
@@ -179,7 +122,7 @@ namespace Ztm.Zcoin.Rpc.Tests
             var tx = Transaction.Create(Network);
 
             return Assert.ThrowsAsync<RPCException>(
-                () => FakeSubject.PopulateExodusInformationAsync(tx, CancellationToken.None)
+                () => Subject.PopulateExodusInformationAsync(tx, CancellationToken.None)
             );
         }
 
@@ -208,7 +151,7 @@ namespace Ztm.Zcoin.Rpc.Tests
             // Act.
             await Assert.ThrowsAsync<ArgumentException>(
                 "tx",
-                () => FakeSubject.PopulateExodusInformationAsync(tx, CancellationToken.None)
+                () => Subject.PopulateExodusInformationAsync(tx, CancellationToken.None)
             );
         }
 
@@ -228,8 +171,8 @@ namespace Ztm.Zcoin.Rpc.Tests
             var transfer = block.Transactions[1];
 
             // Act.
-            await FakeSubject.PopulateExodusInformationAsync(coinbase, CancellationToken.None);
-            await FakeSubject.PopulateExodusInformationAsync(transfer, CancellationToken.None);
+            await Subject.PopulateExodusInformationAsync(coinbase, CancellationToken.None);
+            await Subject.PopulateExodusInformationAsync(transfer, CancellationToken.None);
 
             // Assert.
             Assert.Null(coinbase.GetExodusTransaction());
@@ -251,7 +194,7 @@ namespace Ztm.Zcoin.Rpc.Tests
             Node.Generate(1);
 
             // Act.
-            await FakeSubject.PopulateExodusInformationAsync(tx, CancellationToken.None);
+            await Subject.PopulateExodusInformationAsync(tx, CancellationToken.None);
 
             // Assert.
             Assert.Null(tx.GetExodusTransaction());
@@ -281,7 +224,7 @@ namespace Ztm.Zcoin.Rpc.Tests
             Node.Generate(1);
 
             // Act.
-            await FakeSubject.PopulateExodusInformationAsync(tx, CancellationToken.None);
+            await Subject.PopulateExodusInformationAsync(tx, CancellationToken.None);
 
             // Assert.
             var exodus = (SimpleSendV0)tx.GetExodusTransaction();
@@ -292,109 +235,9 @@ namespace Ztm.Zcoin.Rpc.Tests
             Assert.Equal(owner, exodus.Sender);
         }
 
-        protected virtual RpcClient CreateSubject()
+        protected override RpcClient CreateSubject()
         {
             return new FakeRpcClient(Factory, Client);
-        }
-
-        protected virtual void Dispose(bool disposing)
-        {
-            if (this.disposed)
-            {
-                return;
-            }
-
-            if (disposing)
-            {
-                if (this.subject.IsValueCreated)
-                {
-                    this.subject.Value.Dispose();
-                }
-
-                this.nodes.Dispose();
-            }
-
-            this.disposed = true;
-        }
-
-        protected async Task<uint256> FundAddressAsync(BitcoinAddress address, Money amount)
-        {
-            using (var rpc = await Factory.CreateWalletRpcAsync(CancellationToken.None))
-            {
-                return await rpc.SendAsync(address, amount, null, null, false, CancellationToken.None);
-            }
-        }
-
-        protected async Task<BitcoinAddress> GenerateNewAddressAsync()
-        {
-            using (var rpc = await Factory.CreateWalletRpcAsync(CancellationToken.None))
-            {
-                return await rpc.GetNewAddressAsync(CancellationToken.None);
-            }
-        }
-
-        protected async Task<Block> GetBlockAsync(uint256 hash)
-        {
-            using (var rpc = await Factory.CreateChainInformationRpcAsync(CancellationToken.None))
-            {
-                return await rpc.GetBlockAsync(hash, CancellationToken.None);
-            }
-        }
-
-        protected async Task<Property> GetPropertyAsync(string name)
-        {
-            using (var rpc = await Factory.CreateExodusInformationRpcAsync(CancellationToken.None))
-            {
-                var props = await rpc.ListPropertiesAsync(CancellationToken.None);
-                var prop = props.Single(p => p.Name == name);
-
-                return new Property(prop.Id, prop.Type);
-            }
-        }
-
-        protected async Task<uint256> GrantTokensAsync(
-            Property property,
-            BitcoinAddress from,
-            BitcoinAddress to,
-            PropertyAmount amount)
-        {
-            Transaction tx;
-
-            using (var rpc = await Factory.CreatePropertyManagementRpcAsync(CancellationToken.None))
-            {
-                tx = await rpc.GrantAsync(property, from, to, amount, null, CancellationToken.None);
-            }
-
-            return await SendTransactionAsync(tx);
-        }
-
-        protected async Task<Transaction> SendTokensAsync(
-            Property property,
-            BitcoinAddress from,
-            BitcoinAddress to,
-            PropertyAmount amount)
-        {
-            Transaction tx;
-
-            using (var rpc = await Factory.CreatePropertyManagementRpcAsync(CancellationToken.None))
-            {
-                tx = await rpc.SendAsync(from, to, property, amount, null, CancellationToken.None);
-            }
-
-            using (var rpc = await Factory.CreateRawTransactionRpcAsync(CancellationToken.None))
-            {
-                await rpc.SendAsync(tx, CancellationToken.None);
-            }
-
-            return tx;
-        }
-
-        protected async Task<uint256> SendTransactionAsync(Transaction tx)
-        {
-            using (var rpc = await Factory.CreateRawTransactionRpcAsync(CancellationToken.None))
-            {
-                return await rpc.SendAsync(tx, CancellationToken.None);
-            }
         }
     }
 }
